pyproject.toml
[project]
name = "vectorsdb-homework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "huggingface-hub[hf-xet]>=0.30.2",
    "ipywidgets>=8.1.6",
    "joblib>=1.5.0",
    "matplotlib>=3.10.1",
    "notebook>=7.4.1",
    "pandas>=2.2.3",
    "pgvector>=0.4.1",
    "pillow>=11.2.1",
    "pre-commit>=4.2.0",
    "psycopg>=3.2.7",
    "requests>=2.32.3",
    "sentence-transformers>=4.1.0",
    "sqlalchemy>=2.0.40",
    "torch>=2.7.0",
    "tqdm>=4.67.1",
]
